Extract Race Calendar and its Loader to dedicated classes
