+++ app/src/androidTest/java/com/github/fo2rist/mclaren/tests/calendar/CalendarEventsLoaderTest.java
@@ -4,9 +4,9 @@ import android.content.Context;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
+import com.github.fo2rist.mclaren.ui.calendar.CalendarEventsLoaderImpl;
+import com.github.fo2rist.mclaren.ui.calendar.RaceCalendar;
 import com.github.fo2rist.mclaren.ui.models.CalendarEvent;
-import com.github.fo2rist.mclaren.ui.calendar.CalendarEventsLoader;
-import java.util.List;
 import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Test;
@@ -28,7 +28,7 @@ public class CalendarEventsLoaderTest {
 
     @Test
     public void testCalendar2017() throws Exception {
-        List<CalendarEvent> calendarEvents = new CalendarEventsLoader(context, 2017).getCalendar();
+        RaceCalendar calendarEvents = new CalendarEventsLoaderImpl(context).loadCalendar(2017);
         assertEquals(20, calendarEvents.size());
 
         assertDataCorrect(
@@ -74,7 +74,7 @@ public class CalendarEventsLoaderTest {
 
     @Test
     public void testCalendar2018() throws Exception {
-        List<CalendarEvent> calendarEvents = new CalendarEventsLoader(context, 2018).getCalendar();
+        RaceCalendar calendarEvents = new CalendarEventsLoaderImpl(context).loadCalendar(2018);
         assertEquals(21, calendarEvents.size());
 
         assertDataCorrect(
