+++ app/src/main/java/com/github/fo2rist/mclaren/ui/calendar/RaceCalendar.java
@@ -1,0 +1,7 @@
+package com.github.fo2rist.mclaren.ui.calendar;
+
+import com.github.fo2rist.mclaren.ui.models.CalendarEvent;
+import java.util.ArrayList;
+
+public class RaceCalendar extends ArrayList<CalendarEvent> {
+}
